task and loyalty tasks are added by admin

+++ src/controller/loyaltyTask.ts
@@ -1,5 +1,6 @@
 import { Request, Response } from "express";
 import { LoyaltyTask } from "../database/models/loyaltyTasks";
+import { IUserModel, User } from "../database/models/user";
 
 export const allLoyaltyTask = async (req: Request, res: Response) => {
   const loyaltyTasks = await LoyaltyTask.find();
@@ -7,24 +8,39 @@ export const allLoyaltyTask = async (req: Request, res: Response) => {
 };
 
 export const addLoyaltyTask = async (req: Request, res: Response) => {
-  const checkLoyaltyTask = await LoyaltyTask.findOne({ name: req.body.name });
-  if (!checkLoyaltyTask) {
-    const newLoyaltyTask = new LoyaltyTask({
-      name: req.body.name,
-      type: req.body.type,
-      instruction: req.body.instruction,
-      weight: req.body.weight,
-    });
-    await newLoyaltyTask.save();
-    res.send("loyalty task added");
+  const user = req.user as IUserModel;
+  const userDetails = await User.findOne({ _id: user.id, isModerator: true });
+
+  if (userDetails) {
+    const checkLoyaltyTask = await LoyaltyTask.findOne({ name: req.body.name });
+    if (!checkLoyaltyTask) {
+      const newLoyaltyTask = new LoyaltyTask({
+        name: req.body.name,
+        type: req.body.type,
+        instruction: req.body.instruction,
+        weight: req.body.weight,
+      });
+      await newLoyaltyTask.save();
+      res.send("loyalty task added");
+    } else {
+      res.send("already added");
+    }
+  } else {
+    res.send("not authorized");
   }
-  res.send("already added");
 };
 
 export const deleteLoyaltyTask = async (req: Request, res: Response) => {
-  const checkLoyaltyTask = await LoyaltyTask.findOne({ name: req.body.name });
-  if (checkLoyaltyTask) {
-    LoyaltyTask.deleteOne({ _id: checkLoyaltyTask._id });
-    res.send("loyalty task deleted");
+  const user = req.user as IUserModel;
+  const userDetails = await User.findOne({ _id: user.id, isModerator: true });
+
+  if (userDetails) {
+    const checkLoyaltyTask = await LoyaltyTask.findOne({ name: req.body.name });
+    if (checkLoyaltyTask) {
+      LoyaltyTask.deleteOne({ _id: checkLoyaltyTask._id });
+      res.send("loyalty task deleted");
+    }
+  } else {
+    res.send("not authorized");
   }
 };
